[WIP] Use CTest/CDash for CI builds
DEMONSTRATION ONLY: DO NOT MERGE

This demonstrates using an experimental version of CMake to perform CI
builds via CTest and submit the results to CDash, which provides
additional features, particularly enabled by having a history of prior
builds.

This has historically been rejected due to not having full step output
available through the Github CI logs. The experimental CMake version
used allows individual steps to request verbose output, which hopefully
resolves this objection.

If this approach is acceptable to the community, the experimental
feature can be polished and added to CMake proper.

The CMake binary and CDash dashboard used here are for demonstration
purposes only, and should not be used by the master branch. They may
cease to exist without notice.